app/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
motor==3.3.1
python-multipart>=0.0.9
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.26.0
python-docx>=1.0.0
PyPDF2>=3.0.0